Implement mongo wire protocol decoder for a subset of message opcodes
Based on the information available at mongo's documentation (e.g.
https://docs.mongodb.com/manual/reference/mongodb-wire-protocol/), this
commit introduces a simple decoder for the following opcodes:
- insert
- update
- delete
- getMore
- killCursors

A captured request can be decoded via a call to protocol.Decode. If the
request opcode is not recognized, the decoder will wrap it in an
UnknownRequest and return it as such.