Add scaffolding for a dumb mongo server emulator with pluggable backends
The server is configured with an optional backend to handle incoming CRUD
requests and commands. The server also defines an (extensible) list of
common mongo commands that it can handle itself if the backend cannot.
This approach will let us extract the common command logic (e.g.
isMaster, getNonce, getLastError) and allow us to iterate faster on the
actual backend code.